moves out endpoint handlers and fouresquare service into their own file
